Introduce utility function to calculate inputs to rolling_window
Both polars and pandas specify rolling window extents by providing a
scalar window length, along with whether the interval is open, closed,
left-half-open, or right-half-open, and (for polars) a scalar offset.
We currently use a numba kernel to convert this information into the
pair of `preceding_window` and `following_window` columns that are the
arguments to `rolling_window`. This kernel has a few problems:

- For large window extents it has poor performance scaling as O(n *
  window_size);
- It doesn't support different end point handling;
- Since pandas doesn't require it, it doesn't support computing
  `following_window` information;
- If we want to use it in cudf-polars, it introduces another
  dependency (numba).

To address these concerns, implement a utility function in libcudf
that takes window length and offset and computes the matching
preceding/following columns. This implementation exploits the fact
that the column we are rolling over must be sorted in ascending order
and uses `thrust::lower_bound` to provide an O(n log n) run time for
all window sizes. We also now support all endpoint handling required
for cudf-polars.

This provides the infrastructure to work on addressing (at least
partially):

- #12774
- #14334
- #15086
- #15119
- #15192